Refactor handleInlineShortcuts
